% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPval.R
\name{calcSplitVotPval}
\alias{calcSplitVotPval}
\title{A Function to return raw p-values for feature contribution to a split}
\usage{
calcSplitVotPval(featureMat, labels, queryLabs)
}
\arguments{
\item{featureMat}{Unnamed matrix of observationsxfeature}

\item{labels}{Vector of observation labels (e.g. member names)}

\item{queryLabs}{Subset of labels vector on one side of split}
}
\value{
P-values for contribution of feature to split
}
\description{
A Function to return raw p-values for feature contribution to a split
}
