% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPval.R
\name{calcVotPval}
\alias{calcVotPval}
\title{Get raw p-values for feature contribution to a split of interest}
\usage{
calcVotPval(featureMat, labels, queryLabs)
}
\arguments{
\item{featureMat}{Unnamed matrix of observationsxfeatures (e.g. membersxvotes)}

\item{labels}{Vector of observation labels (e.g. member names)}

\item{queryLabs}{Subset of labels vector on one side of split}
}
\value{
P-values for contribution of feature to split
}
\description{
Get raw p-values for feature contribution to a split of interest
}
\examples{
featMat <- cbind(c(0,1,1,1,1),c(0,0,0,1,1),c(0,1,1,1,1),c(0,1,0,1,1),c(0,0,0,0,0),c(0,1,1,0,0))
labs <- c('Mouse','Fly','Rat','Zebrafish','Human')
queries <- c('Mouse','Rat')
pvals <- calcVotPval(featMat,labs,queries)
}
