% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dists}
\alias{dists}
\title{116th Senate Pairwise Distances}
\format{
A data frame with 98 rows and 2 variables:
\describe{
\item{Dists}{Matrix of pairwise distances}
\item{Labs}{Row labels for each Senator}
}
}
\source{
\url{https://www.github.com/tarachari3/SplitStats}
}
\usage{
dists
}
\description{
A data set with a matrix for pairwise distances between Senators in the 116th Senate and a list of label names for each Senator
}
\keyword{datasets}
