% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPval.R
\name{getSplitDist}
\alias{getSplitDist}
\title{Get split distances between two groups of observations (members)}
\usage{
getSplitDist(labels, queryLabs, refLabs, splits)
}
\arguments{
\item{labels}{: Vector of observation labels}

\item{queryLabs}{: Vector of observation labels on one side of a split}

\item{refLabs}{: Vector of reference observations labels}

\item{splits}{: Collection of split objects generated by getSplits}
}
\value{
Vector of distances from each query input to set of all observations in refLabs
}
\description{
Get split distances between two groups of observations (members)
}
\examples{
dists <- dist(cbind(c(0,1,1,1,1),c(0,0,0,1,1),c(0,1,1,1,1),c(0,1,0,1,1),c(0,0,0,0,0),c(0,1,1,0,0)), method = "manhattan",diag=TRUE,upper=TRUE)
labs <- c('Mouse','Fly','Rat','Zebrafish','Human')
out <- getSplits(labs, dists)
splits <- out[2][[1]]
sdists <- getSplitDist(labs, labs[1:3], labs[4:6], splits)
}
