% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPval.R
\name{getSplits}
\alias{getSplits}
\title{Get splits and cycle ordering from NNet generated split network}
\usage{
getSplits(labels, matrix)
}
\arguments{
\item{labels}{Vector of observation labels (e.g. member names)}

\item{matrix}{(Square) Distance matrix with no labels (unnamed)}
}
\value{
List of observations in cycle ordering and split objects
}
\description{
Get splits and cycle ordering from NNet generated split network
}
