% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPval.R
\name{getSplitVis}
\alias{getSplitVis}
\title{Visualize split network as splits graph}
\usage{
getSplitVis(labels, cycle, splits, matrix, outfilePhylo, outfileNexus)
}
\arguments{
\item{labels}{: Vector of labels for members (observations) in dataset}

\item{cycle}{: Vector of circular ordering of labels from getSplits}

\item{splits}{: Split objects between members from getSplits}

\item{matrix}{: Square distance matrix (used as input to getSplits)}

\item{outfilePhylo}{: Filename for phylogenetic outline image of splits network}

\item{outfileNexus}{: Filename for NEXUS output of splits network (SplitsTree4/5 compatible)}
}
\value{
Image and NEXUS file output
}
\description{
Visualize split network as splits graph
}
\author{
Tara Chari
}
