% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{votes}
\alias{votes}
\title{116th Senate Votes}
\format{
A matrix with 98 members by 168 rollcall votes:
\describe{
\item{1}{Matrix of Senate member votes}
}
}
\source{
\url{https://www.github.com/tarachari3/SplitStats}
}
\usage{
votes
}
\description{
A data set with a matrix for Senator Rollcall votes (0,1/2,1) for (Nay,Abstain,Yea)
}
\keyword{datasets}
