% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitTools.R
\name{getSplitDist}
\alias{getSplitDist}
\title{Get split distances between two groups of observations (e.g. members)}
\usage{
getSplitDist(labels, queryLabs, refLabs, splits)
}
\arguments{
\item{labels}{: Vector of observation labels}

\item{queryLabs}{: Vector of observation labels on one side of a split (subset of labels)}

\item{refLabs}{: Vector of reference observations labels (subset of labels)}

\item{splits}{: Collection of split objects generated by getSplits}
}
\value{
Vector of distances from each query input to set of all observations in refLabs
}
\description{
Get split distances between two groups of observations (e.g. members)
}
\examples{
featMat <- cbind(c(0,1,1,1,1),c(0,0,0,1,1),c(0,1,1,1,1),c(0,1,0,1,1),c(0,0,0,0,0),c(0,1,1,0,0))
ds <- dist(featMat, method = "manhattan",diag=TRUE,upper=TRUE)
ds <- as.matrix(ds)
labs <- c('Mouse','Fly','Rat','Zebrafish','Human')
out <- getSplits(labs, ds)
splits <- out[2][[1]]
sdists <- getSplitDist(labs, labs[1:3], labs[4:6], splits)
}
