% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPval.R
\name{getSplits}
\alias{getSplits}
\title{Get splits and cycle ordering from NNet generated split network}
\usage{
getSplits(labels, matrix)
}
\arguments{
\item{labels}{Vector of observation labels (e.g. member names)}

\item{matrix}{(Square) Distance matrix with no labels (unnamed)}
}
\value{
List of observations in cycle ordering and split objects
}
\description{
Get splits and cycle ordering from NNet generated split network
}
\examples{
dists <- dist(cbind(c(0,1,1,1,1),c(0,0,0,1,1),c(0,1,1,1,1),c(0,1,0,1,1),c(0,0,0,0,0),c(0,1,1,0,0)), method = "manhattan",diag=TRUE,upper=TRUE)
labs <- c('Mouse','Fly','Rat','Zebrafish','Human')
out <- getSplits(labs, dists)
}
