% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPval.R
\name{getSplitVis}
\alias{getSplitVis}
\title{Visualize split network as splits graph}
\usage{
getSplitVis(
  labels,
  cycle,
  splits,
  matrix,
  outfilePhylo,
  outfileNexus,
  show = TRUE,
  width = 1000,
  height = 800,
  m_left = 100,
  m_right = 100,
  m_top = 100,
  m_bot = 100,
  font_size = 12,
  scale_factor = 5
)
}
\arguments{
\item{labels}{: Vector of labels for members (observations) in dataset}

\item{cycle}{: Vector of circular ordering of labels from getSplits}

\item{splits}{: Split objects between members from getSplits}

\item{matrix}{: Square distance matrix (used as input to getSplits)}

\item{outfilePhylo}{: Filename for phylogenetic outline image of splits network}

\item{outfileNexus}{: Filename for NEXUS output of splits network (SplitsTree4/5 compatible)}

\item{show}{: Boolean for showing outline on console}

\item{width}{: Width int}

\item{height}{: Height int}

\item{m_left}{: Restrict width on left by int}

\item{m_right}{: Restrict width on right by int}

\item{m_top}{: Restrict height on top by int}

\item{m_bot}{: Restrict height on bottom by int}

\item{font_size}{: Font size}

\item{scale_factor}{: Scaling outline for image}
}
\value{
Image and NEXUS file output
}
\description{
Visualize split network as splits graph
}
\examples{
dists <- dist(cbind(c(0,1,1,1,1),c(0,0,0,1,1),c(0,1,1,1,1),c(0,1,0,1,1),c(0,0,0,0,0),c(0,1,1,0,0)), method = "manhattan",diag=TRUE,upper=TRUE)
labs <- c('Mouse','Fly','Rat','Zebrafish','Human')
out <- getSplits(labs, dists)
splits <- out[2][[1]]
cycle <- out[1][[1]]
getSplitVis(labs,cycle,splits,dists,'','')
}
\author{
Tara Chari
}
